backend/requirements.txt
django-cors-headers==2.2.0
django-filter==1.1.0
django-phonenumber-field==2.0.0
Django==2.0.7
djangorestframework==3.8.2
drf-nested-routers==0.90.2
expects==0.8.0
factory_boy==2.11.1
faker-e164==0.1.0
gunicorn==19.9.0
mamba==0.9.3
markdown==2.6.11
psycopg2==2.7.5
